+++ Milles_Project1Library/Models/Calculator.cs
@@ -12,12 +12,12 @@ namespace Milles_Project1Library.Models
         [Key]
         public int CalculationId { get; set; }
 
-        public double Number1 { get; set; }
-        public double Number2 { get; set; }
+        public decimal Number1 { get; set; }
+        public decimal Number2 { get; set; }
 
         public string Operator { get; set; }
 
-        public double Result { get; set; }
+        public decimal Result { get; set; }
 
         public DateTime CalculationDate { get; set; }
     }
